requirements.txt
flask
opencv-python
Pillow
google-generativeai
python-dotenv   # for managing your GEMINI_API_KEY in .env
gunicorn        # if you plan to deploy
